+++ todo/src/main/java/com/todoback/todo/services/UserService.java
@@ -7,9 +7,9 @@ import com.todoback.todo.dto.UserDto;
 public interface UserService {
 	
 	UserDto save(UserDto user);
-	void delete(int id);
+	void delete(Long id);
 	List<UserDto> findAll();
-	UserDto findById(int id);
+	UserDto findById(Long id);
 	UserDto login(UserDto user);
 	
 }
